Transferring unittest functions to test_base.py
