+++ tests/vasp/test_vasp.py
@@ -27,26 +27,6 @@ class TestVasp(unittest.TestCase):
         self.assertIsInstance(self.job.input, Input)
         self.assertIsInstance(self.job._output_parser, Output)
 
-    def test_calc_static(self):
-        self.job.calc_static(electronic_steps=90, retain_charge_density=True, retain_electrostatic_potential=True)
-        self.assertEqual(self.job.input.incar["IBRION"], -1)
-        self.assertEqual(self.job.input.incar["NELM"], 90)
-        self.assertEqual(self.job.input.incar["LVTOT"], True)
-        self.assertEqual(self.job.input.incar["LCHARG"], True)
-
-    def test_set_structure(self):
-        self.assertEqual(self.job.structure, None)
-        atoms = CrystalStructure("Pt", BravaisBasis="fcc", a=3.98)
-        self.job.structure = atoms
-        self.assertEqual(self.job.structure, atoms)
-        self.job.structure = None
-        self.assertEqual(self.job.structure, None)
-        self.job.structure = atoms
-        self.assertEqual(self.job.structure, atoms)
-
-    def test_list_potenitals(self):
-        self.assertRaises(ValueError, self.job.list_potentials)
-
     def tearDown(self):
         pass
